pipelines: pin actions/setup-go action to commit hash (#42)
While we are still waiting for official approval for the `setup-go`
action, we should still pin the action to a specific commit hash in
preparation for approval. Currently, the v5 tag points to v5.3.0, so
this is a version downgrade, but this is the commit we asked labs
for approval.